model-controller advanced frame
