--- core/src/main/java/io/github/mole/controll/GameController.java
@@ -1,4 +1,0 @@
-package io.github.mole.controll;
-
-public class GameController {
-}
