+++ core/src/main/java/io/github/mole/controllerr/GameController.java
@@ -1,0 +1,4 @@
+package io.github.mole.controllerr;
+
+public class GameController {
+}
