+++ core/src/main/java/io/github/mole/controllerr/attacks/WaterAttack.java
@@ -1,0 +1,4 @@
+package io.github.mole.controllerr.attacks;
+
+public class WaterAttack {
+}
