+++ core/src/main/java/io/github/mole/controllerr/environment/MossController.java
@@ -1,0 +1,4 @@
+package io.github.mole.controllerr.environment;
+
+public class MossController {
+}
