+++ core/src/main/java/io/github/mole/controllerr/main/HumanController.java
@@ -1,0 +1,4 @@
+package io.github.mole.controllerr.main;
+
+public class HumanController {
+}
