+++ core/src/main/java/io/github/mole/controllerr/main/MoleController.java
@@ -1,0 +1,4 @@
+package io.github.mole.controllerr.main;
+
+public class MoleController {
+}
