--- core/src/main/java/io/github/mole/mod/Board.java
@@ -1,4 +1,0 @@
-package io.github.mole.mod;
-
-public class Board {
-}
