+++ core/src/main/java/io/github/mole/mod/Model.java
@@ -1,5 +1,8 @@
 package io.github.mole.mod;
 
+import io.github.mole.mod.main.Board;
+import io.github.mole.mod.main.Mole;
+
 public class Model {
     Board board;
     Mole mole;
