+++ core/src/main/java/io/github/mole/mod/attacks/PetardAttack.java
@@ -1,0 +1,4 @@
+package io.github.mole.mod.attacks;
+
+public class PetardAttack {
+}
