+++ core/src/main/java/io/github/mole/mod/attacks/WaterAttack.java
@@ -1,0 +1,4 @@
+package io.github.mole.mod.attacks;
+
+public class WaterAttack {
+}
