+++ core/src/main/java/io/github/mole/mod/environment/Moss.java
@@ -1,0 +1,4 @@
+package io.github.mole.mod.environment;
+
+public class Moss {
+}
