+++ core/src/main/java/io/github/mole/mod/main/Human.java
@@ -1,0 +1,4 @@
+package io.github.mole.mod.main;
+
+public class Human {
+}
