+++ core/src/main/java/io/github/mole/mod/main/Mole.java
@@ -1,0 +1,4 @@
+package io.github.mole.mod.main;
+
+public class Mole {
+}
