--- core/src/main/java/io/github/mole/mod/specialities/Spade.java
@@ -1,4 +1,0 @@
-package io.github.mole.mod.specialities;
-
-public class Spade {
-}
